+++ lib/ssr/deno.rb
@@ -4,6 +4,7 @@ require_relative 'deno/version'
 
 # Load the native Rust extension (compiled by rb-sys / rake-compiler)
 require_relative 'deno/ssr_deno'
+require_relative 'deno/render_error'
 require_relative 'deno/heap_stats'
 require_relative 'deno/config'
 require_relative 'deno/instrumenter'